third_party/WebKit/LayoutTests/http/tests/devtools/oopif/oopif-elements-nesting-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that oopif iframes are rendered inline.


Running: testSetUp

Running: testRemove
- <html>
      <head></head>
    - <body>
        - <div>
            - <iframe id="page-iframe1" src="http://127.0.0.1:8000/devtools/oopif/resources/iframe.html?first">
                - #document
                    - <html>
                          <head></head>
                        - <body>
                            - <div>
                                - <iframe id="iframe-iframe1" src="http://127.0.0.1:8000/devtools/oopif/resources/inner-iframe.html?first">
                                    - #document
                                        - <html>
                                              <head></head>
                                              <body>IFRAME\n</body>
                                          </html>
                                  </iframe>
                              </div>
                            - <div>
                                - <iframe id="iframe-iframe2" src="http://devtools.oopif.test:8000/devtools/oopif/resources/inner-iframe.html?second">
                                    - #document
                                        - <html>
                                              <head></head>
                                              <body>IFRAME\n</body>
                                          </html>
                                  </iframe>
                              </div>
                          </body>
                      </html>
              </iframe>
          </div>
        - <div>
            - <iframe id="page-iframe2" src="http://devtools.oopif.test:8000/devtools/oopif/resources/iframe.html?second">
                - #document
                    - <html>
                          <head></head>
                        - <body>
                            - <div>
                                - <iframe id="iframe-iframe1" src="http://127.0.0.1:8000/devtools/oopif/resources/inner-iframe.html?first">
                                    - #document
                                        - <html>
                                              <head></head>
                                              <body>IFRAME\n</body>
                                          </html>
                                  </iframe>
                              </div>
                            - <div>
                                - <iframe id="iframe-iframe2" src="http://devtools.oopif.test:8000/devtools/oopif/resources/inner-iframe.html?second">
                                    - #document
                                        - <html>
                                              <head></head>
                                              <body>IFRAME\n</body>
                                          </html>
                                  </iframe>
                              </div>
                          </body>
                      </html>
              </iframe>
          </div>
      </body>
  </html>
